Fix realtime staff attendances tests
